updated abstract classes to use Integer for ID. this is specific to the ERA scheam table business_unit_catalog. it is defined as a Integer instead of Long.  we still need to push for their db admin to create an auto incrementing index on that table. other wise, api insert does not work.

+++ common/src/main/java/gov/nara/common/interfaces/IOperations.java
@@ -7,7 +7,7 @@ public interface IOperations<T extends Serializable> {
 
     // find - one
 
-    T findOne(final long id);
+    T findOne(final Integer id);
 
     /**
      * - contract: if nothing is found, an empty list will be returned to the calling client <br>
@@ -39,7 +39,7 @@ public interface IOperations<T extends Serializable> {
 
     // delete
 
-    void delete(final long id);
+    void delete(final Integer id);
 
     void deleteAll();
